Add reindex embeddings command alias
