+++ plugins/remem/runtimes/remem-releases.json
@@ -1,7 +1,7 @@
 {
   "versions": {
-    "0.5.56": {
-      "base_url": "https://github.com/majiayu000/remem/releases/download/v0.5.56",
+    "0.5.57": {
+      "base_url": "https://github.com/majiayu000/remem/releases/download/v0.5.57",
       "assets": {}
     }
   }
